[Ripple][LLVM] Vector For Loop: Fix LoopIncBBs
In computing LoopIncBBs, we need all BBs that are the on the path
leading to the back-edge to BranchingBB. During this computation,
we have to explicitly ignore BBs that are on the loop cleanup path.

For example, consider the CFG below:

```
┌─ A ─┐◄───┐
│     │    │
│     ▼    │
│     B ───┘
│     ▼
└──►  C ─► D
```
. Our handling of back edges with vector conditions needs the
BBs that correspond to the loop increment path i.e. the ones
on the path leading to the back-edge. So, in this case, we need
just "B". Consequently, when performing the traversal to find
these basic blocks we must traverse the successors while ignore
the basic blocks that correspond to the loop cleanup phase (i.e.
'C').

This patch includes the fix for ignoring the cleanup BBs during
the above mentioned traversal. A test 9vecforloop.ll is also
attached that would cause an assertion error without this patch.

Reviewed-by: Benoît Meister <bmeister@qti.qualcomm.com>