[Ripple][LLVM] LinearSeries better cast propagation
We allow the propagation of LinearSeries through cast when the semantics
is preserved. That means that we can only propagate a linear series
through a cast when no-wrap semantics allows it:

- Allow LinearSeries to propagate through Truncate(X) iff Trunc(X) (and
all the succesors through PHI and Select) have no-wrap semantics.
- Allow LinearSeries to propagate through Extend(X) iff X (or all
predecessors if X is a PHI or Select) have no-wrap semantics.

Validity proof:

For truncate: by considering Truncate(AX + B) as Truncate(AX) +
Truncate(B) in subsequent <OP>s, and all these <OP>s have no-wrap
semantics, we are guaranteed that if a value overflows, we hit an
undefined behavior allowing us to propagate.

For extend: by considering Extend(AX + B) as Extend(AX) + Extend(B), and
(AX + B) has no-wrap semantics, we are guaranteed that we already hit an
undefined behavior before the Extend on overflows, allowing us to
propagate.

Reviewed-by: Benoit Meister <benoit.meister@qti.qualcomm.com>