[Ripple][LLVM] LinearSeries propagate broad-casts
Casts block the propagation of LinearSeries unless they meet specific
no-wrap conditions. However, if a LinearSeries represents a broadcast
(all slopes are zero) or a scalar, the cast operation applies uniformly
to the base value, and the "slopes" remain zero.

This patch allows propagating such "zero-slope" LinearSeries through any
CastInst by casting the base and maintaining zero slopes. This enables
further optimizations where the broadcasted value is used.

Similarly, binary operators allow the propagation of broadcast-only
slopes when the base shape of the LHS and RHS are equal.

Unary operators allow the propagation of broadcasts.

Reviewed-by: Benoit Meister <benoit.meister@qti.qualcomm.com>